+++ src/http/modules/ngx_http_proxy_module.c
@@ -2343,7 +2343,7 @@ ngx_http_proxy_lowat_check(ngx_conf_t *cf, void *post, void *data)
 #if (NGX_FREEBSD)
     ssize_t *np = data;
 
-    if (*np >= ngx_freebsd_net_inet_tcp_sendspace) {
+    if ((u_long) *np >= ngx_freebsd_net_inet_tcp_sendspace) {
         ngx_conf_log_error(NGX_LOG_EMERG, cf, 0,
                            "\"proxy_send_lowat\" must be less than %d "
                            "(sysctl net.inet.tcp.sendspace)",
